Corrigindo

+++ Server/rosetreferencia/main.py
@@ -28,7 +28,7 @@ def on_request(ch, method, properties, body):
 		for i in range(0,int(num)):
 			mycursor = mydb.cursor()
 			comand = "UPDATE Referencia SET (Preco,Tempo) = (%s,%s) WHERE Nome = %s"
-			val = (float(ref[3*i+1]),int(ref[3*i+2]),ref[3*i])
+			val = (ref[3*i+1],ref[3*i+2],ref[3*i])
 			mycursor.execute(comand,val)
 			mydb.commit()
